% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_prep.R
\name{add_bottom_row}
\alias{add_bottom_row}
\title{Add Bottom Row}
\usage{
add_bottom_row(data, value = NA_real_)
}
\arguments{
\item{data}{}
}
\description{
Adds an empty row to the bottom of a \code{dataframe}
}
