% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_prep.R
\name{add_parentheses}
\alias{add_parentheses}
\title{Add Parentheses}
\usage{
add_parentheses(vector)
}
\arguments{
\item{vector}{}
}
\description{
Adds parenthesis to input
}
\examples{
add_parentheses(45)
(45)

}
