% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{beta_2_d}
\alias{beta_2_d}
\title{Regression Coefficient To D}
\usage{
beta_2_d(beta, standard_error, n)
}
\arguments{
\item{beta}{Beta from regression}

\item{standard_error}{Corrosponding standard error}

\item{n}{Corrosponding sample size}
}
\description{
Regression Coefficient To D
}
