% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{beta_2_r}
\alias{beta_2_r}
\title{Regression Coefficient To R}
\usage{
beta_2_r(beta, standard_error, n)
}
\arguments{
\item{beta}{Beta from regression}

\item{standard_error}{Corrosponding standard error}

\item{n}{Corrosponding sample size}
}
\value{
r value
}
\description{
Regression Coefficient To R
}
