% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing_functions.R
\name{canonical_correlation}
\alias{canonical_correlation}
\title{Canonical Correlations}
\usage{
canonical_correlation(data, dv)
}
\arguments{
\item{dv}{}
}
\description{
Canonical Correlations
}
