% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{ci_2_sd}
\alias{ci_2_sd}
\title{Confidence Interval to Standard Deviation}
\usage{
ci_2_sd(upper_ci, lower_ci, n, interval = 95)
}
\arguments{
\item{upper_ci}{Upper confidence interval}

\item{lower_ci}{Lower confidence interval}

\item{n}{Sample size}

\item{interval}{Either 95, or 90. 95 by default.}
}
\description{
Confidence Interval to Standard Deviation
}
