% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{CI_converter}
\alias{CI_converter}
\title{CI converter}
\usage{
CI_converter(upper, lower, n, interval = 95)
}
\arguments{
\item{upper}{Upper confidence interval}

\item{lower}{Lower confidence interval}

\item{n}{Sample size}

\item{interval}{Either 95, or 90. 95 by default.}
}
\description{
CI converter
}
