% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing_functions.R
\name{corr_multiple_testing}
\alias{corr_multiple_testing}
\title{Correlations for Multiple Testing}
\usage{
corr_multiple_testing(dataset, dv, unadj_p = TRUE, bh = TRUE,
  permutation = TRUE, save_conf_level = FALSE,
  print_permutation_summary = FALSE, conf_level = 0.95, nrep = 100)
}
\arguments{
\item{nrep}{}
}
\description{
Correlations for Multiple Testing
}
