% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_folder.R
\name{data_folder}
\alias{data_folder}
\title{Data Folder}
\usage{
data_folder(dataset = "", data_folder = "", ws = FALSE)
}
\arguments{
\item{dataset}{The name of the dataset to load as string.}

\item{ws}{Should there be white space is preceding or trailing the name of the datafolder? Defaults to FALSE.}

\item{datafolder}{The name where the dataset is stored one directory back. Directories "Data", "data", "dataset", "datasets" are checked automatically.}
}
\description{
A function so that users don't need to change where data is loaded from when multiple users are working on a single project. Most likely nested inside the function read.csv() or read_csv().
}
\examples{
read.csv(data_folder("mydataset.csv"))
}
