% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{did_calculator}
\alias{did_calculator}
\title{Difference in Difference}
\usage{
did_calculator(mean_treatment_post, mean_treatment_pre, mean_control_post,
  mean_control_pre, sd)
}
\arguments{
\item{sd}{Control standard deviation}
}
\description{
Difference in Difference
}
