% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_prep.R
\name{kable_resize}
\alias{kable_resize}
\title{Kable Resize}
\usage{
kable_resize(table, resizebox = 0.8, minipage = 1)
}
\arguments{
\item{minipage}{}
}
\description{
An option for when \code{kable_styling(full_width = F, latex_options = "scale_down")} isn't enough. This function depends on you already attempting to resize using the aforementioned function.
When using with endnotes, I recommend to use this function with \code{kableExtra}'s built in, so that the resize corrosponds with the table
footnote function
}
