% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{merged_r}
\alias{merged_r}
\title{Merged R}
\usage{
merged_r(r, weighted_r)
}
\arguments{
\item{r}{\code{Vector} of pearson's R correlations}

\item{weighted_r}{\code{Vector} of weighted R functions. See \code{weighted_r}}
}
\value{
Single meta-analysis r value
}
\description{
Merged R
}
