% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing_functions.R
\name{multiple_testing}
\alias{multiple_testing}
\title{Correlations for Multiple Testing}
\usage{
multiple_testing(dataset, dv, unadj_p = TRUE, bh = TRUE,
  permutation = TRUE, nrep = 100)
}
\arguments{
\item{nrep}{}
}
\description{
Correlations for Multiple Testing
}
