% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiple_testing_functions.R
\name{partial_multiple_testing}
\alias{partial_multiple_testing}
\title{Partial Correlations for Multiple Testing}
\usage{
partial_multiple_testing(data, dv, controls, unadj_p = TRUE, bh = TRUE,
  permutation = TRUE, nrep = 100)
}
\arguments{
\item{nrep}{}
}
\description{
Partial Correlations for Multiple Testing
}
