% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robust_se.R
\name{robust_se}
\alias{robust_se}
\title{Robust Standard Error}
\usage{
robust_se(model, cluster)
}
\arguments{
\item{cluster}{What variable should the data be clustered by}
}
\description{
Robust Standard Error
}
