% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{sd_pooled}
\alias{sd_pooled}
\title{Standard Deviation Pooled}
\usage{
sd_pooled(sd, n)
}
\arguments{
\item{sd}{Vector of standard deviations}

\item{n}{Vector of sample sizes}
}
\description{
Standard Deviation Pooled
}
