% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{se_2_sd}
\alias{se_2_sd}
\title{Standard Error to Standard Deviation}
\usage{
se_2_sd(se, n)
}
\arguments{
\item{se}{Standard error}

\item{n}{Sample size}
}
\description{
Standard Error to Standard Deviation
}
