% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{stand.result}
\alias{stand.result}
\title{Standardized Result}
\usage{
stand.result(eff.type, u.s.d, ctrl.sd, n.t, n.c)
}
\arguments{
\item{eff.type}{Effect type. Either \code{d.i.d}, \code{d.i.m}, \code{d}, \code{reg.coef}, \code{t.test}, \code{f.test}}

\item{u.s.d}{Unstandardized effect size}

\item{ctrl.sd}{Control standard deviation}

\item{n.t}{Treatment sample size}

\item{n.c}{Control group sample size}
}
\description{
Standardized Result
}
