% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/star_ready.R
\name{star_ready}
\alias{star_ready}
\title{Star Ready
Prepares output from \code{tidy_lm} for use with the package stargazer}
\usage{
star_ready(tidy_lm_df, data)
}
\arguments{
\item{tidy_lm_df}{A \code{tibble} from the \code{tidy_lm} function. No columns can be missing or out of the original order. Rows can be appended so long as they use the same data.}

\item{data}{Dataframe with corresponding data that was used in the tidy_lm function.}
}
\value{
Returns a tibble with all lms that can be directly used with stargazer.
}
\description{
Star Ready
Prepares output from \code{tidy_lm} for use with the package stargazer
}
