% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_prep.R
\name{stargazer_pvalues}
\alias{stargazer_pvalues}
\title{Stargazer P Values}
\usage{
stargazer_pvalues(table, format)
}
\arguments{
\item{table}{Table from \code{stargazer} using the \code{capture.output()} function}

\item{format}{String either 'brackets', 'parentheses' or 'none' to replace the stargazer default P = x}
}
\description{
The default way for stargazer to report P-Values is by "P = X". This function offers several ways to change this.
}
\examples{
model <- lm(hp ~ wt, mtcars)
table_rough <- stargazer(model, report = "csp")
table <- stargazer_pvalues(table_rough, "brackets")
print_table(table)
}
