% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_prep.R
\name{stargazer_rowname}
\alias{stargazer_rowname}
\title{Stargazer Rowname}
\usage{
stargazer_rowname(table, row_name, row_number)
}
\arguments{
\item{row_name}{Name for row}

\item{row_number}{Which row number to add the name to. Best determined by looking at the object created by \code{capture.output()}}
}
\description{
Manually adds rownames to in first column
}
