% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta_analysis.R
\name{t_inverse}
\alias{t_inverse}
\title{T Inverse}
\usage{
t_inverse(p, n)
}
\arguments{
\item{p}{P-value associated with t-test}

\item{n}{Sample size associated with t-test}
}
\description{
Calculates t value from p, and n values
}
