% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_prep.R
\name{table_numbers}
\alias{table_numbers}
\title{Renumber LaTeX Table in a document}
\usage{
table_numbers(latex_table, hold_primary_count = TRUE,
  rm_primary_count = FALSE, rm_secondary_count = FALSE,
  reset_secondary_counter = FALSE, primary_count_style = "arabic",
  secondary_count_style = "alph")
}
\arguments{
\item{hold_primary_count}{Stops the first counter from ascending}

\item{rm_primary_count}{Does not show primary count}

\item{rm_secondary_count}{Removes secondary count}

\item{reset_secondary_counter}{Resets the seconary counter back to 1}

\item{primary_count_style}{\code{Character} vector that should be equal to \code{alph}, \code{Alph}, \code{arabic}, \code{roman}, or \code{Roman}.}

\item{secondary_count_style}{\code{Character} vector that should be equal to \code{alph}, \code{Alph}, \code{arabic}, \code{roman}, or \code{Roman}.}
}
\description{
Requires \code{header needs \newcounter{foo}} to be added in document YAML to work under header-includes:. Useful for adding secondary
counts to tables (e.g., Table 2a, Table 2b). Can also be used to switch primary counts to alternative styles like \code{alph} or \code{roman}.
}
