% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_lm.R
\name{tidy_lm}
\alias{tidy_lm}
\title{tidy_lm}
\usage{
tidy_lm(dv, terms, style = "default", treatment = NULL,
  clusters = NULL, print_summary = FALSE, data = .)
}
\arguments{
\item{dv}{Vector of one or more dependent variables. The function will regress all the terms on each dv separately}

\item{terms}{Vector of terms to regress on dv.}

\item{style}{Character string either "default", "incremental", or "single". In what way should the terms regress on dv. "default" uses all terms at once on all the DVs. "incremental" starts off with one term and adds one more term each model for each dv. "single" uses all terms individually on each dv.}

\item{treatment}{Vector for terms to include in the output dataset. Will return coefficent, p value, and SE.}

\item{clusters}{Clusters for grouping standard error.}

\item{print_summary}{Boolean whether to print model summaries or not. Default is FALSE.}

\item{data}{A data.frame}
}
\value{
A tibble containing all models, and data on what variables are included in each model
}
\description{
A function to streamline the process of running regressions in experimental psychology
}
