% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_prep.R
\name{undo_closure}
\alias{undo_closure}
\title{Undo Closure}
\usage{
undo_closure(data)
}
\arguments{
\item{data}{}
}
\description{
Removes one character from each side of a \code{string}. Useful for undoing functions like \code{add_parentheses} or \code{add_brackets}.
}
\examples{
undo_closure("(45)")
"45"
}
